# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "https://maybets.com/terms",
        "contact": {
            "name": "API Support",
            "url": "https://maybets.com/contact-us",
            "email": "tech@maybets.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/outcome/alias": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This API gets all sms markets alias",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin - markets"
                ],
                "summary": "Get sms markets alias",
                "responses": {
                    "200": {
                        "description": "MarketAlias array ",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.MarketAlias"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid payload or missing field in the payload",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authorization error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Template not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This API creates  sms market alias",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin - markets"
                ],
                "summary": "create sms market alias",
                "parameters": [
                    {
                        "description": "MarketAlias Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateMarketAlias"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": " created successfully",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payload or missing field in the payload",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authorization error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/outcome/alias/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This API updates  sms market alias",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin - markets"
                ],
                "summary": "Update sms market alias",
                "parameters": [
                    {
                        "description": "MarketAlias Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateMarketAlias"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "SMS Alias ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": " created successfully",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payload or missing field in the payload",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authorization error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This API deletes  sms market alias",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin - markets"
                ],
                "summary": "delete sms market alias",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "SMS Alias ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": " deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payload or missing field in the payload",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authorization error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateMarketAlias": {
            "type": "object",
            "properties": {
                "alias": {
                    "description": "Alias custom alias",
                    "type": "string"
                },
                "market_id": {
                    "description": "MarketID market ID",
                    "type": "integer"
                },
                "outcome_id": {
                    "description": "OutcomeID market outcome ID",
                    "type": "string"
                },
                "specifier": {
                    "description": "Specifier market line specifier",
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "required": [
                "error_code",
                "error_message"
            ],
            "properties": {
                "error_code": {
                    "description": "ErrorCode error status code",
                    "type": "integer"
                },
                "error_message": {
                    "description": "ErrorMessage status description",
                    "type": "string"
                }
            }
        },
...
        "models.MarketAlias": {
            "type": "object",
            "properties": {
                "alias": {
                    "description": "Alias custom alias",
                    "type": "string"
                },
                "id": {
                    "description": "ID",
                    "type": "integer"
                },
                "market_id": {
                    "description": "MarketID market ID",
                    "type": "integer"
                },
                "market_name": {
                    "description": "MarketName market name",
                    "type": "string"
                },
                "outcome_id": {
                    "description": "OutcomeID market outcome ID",
                    "type": "string"
                },
                "specifier": {
                    "description": "Specifier market line specifier",
                    "type": "string"
                }
            }
        },
        "models.SuccessResponse": {
            "type": "object",
            "required": [
                "data",
                "status"
            ],
            "properties": {
                "data": {
                    "description": "Message status description",
                    "type": "string"
                },
                "status": {
                    "description": "Status http status code",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "api-key",
            "in": "header"
        }
    }
}